+++ src/Resources/contao/languages/de/tl_settings.php
@@ -1,0 +1,14 @@
+<?php if (!defined('TL_ROOT')) die('You cannot access this file directly!');
+
+/**
+ * legends
+ */
+$GLOBALS['TL_LANG']['tl_settings']['schiedsrichter_legend']       = 'Schiedsrichter';
+
+/**
+ * fields
+ */
+$GLOBALS['TL_LANG']['tl_settings']['schiedsrichter_host'] = array('Host', 'Host-Adresse der Ergebnisdienst-Datenbank');
+$GLOBALS['TL_LANG']['tl_settings']['schiedsrichter_db'] = array('Datenbank', 'Name der Ergebnisdienst-Datenbank');
+$GLOBALS['TL_LANG']['tl_settings']['schiedsrichter_user'] = array('Benutzer', 'Benutzer der Ergebnisdienst-Datenbank');
+$GLOBALS['TL_LANG']['tl_settings']['schiedsrichter_pass'] = array('Passwort', 'Passwort der Ergebnisdienst-Datenbank');